backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
reportlab==4.2.5
python-dateutil==2.9.0
click==8.1.7
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
APScheduler==3.10.4
slowapi==0.1.9
openai>=1.10.0
